refactor: update project stucture
