+++ .env.example
@@ -1,5 +1,5 @@
-REACT_APP_AUTH0_DOMAIN=AUTH0-DOMAIN
-REACT_APP_AUTH0_CLIENT_ID=AUTH0-CLIENT-ID
-REACT_APP_AUTH0_CALLBACK_URL=http://localhost:5173/callback
-REACT_APP_API_SERVER_URL=http://localhost:8000
-REACT_APP_AUTH0_AUDIENCE=AUTH0-AUDIENCE
+VITE_AUTH0_DOMAIN=AUTH0-DOMAIN
+VITE_AUTH0_CLIENT_ID=AUTH0-CLIENT-ID
+VITE_AUTH0_CALLBACK_URL=http://localhost:5173/callback
+VITE_API_SERVER_URL=http://localhost:6060
+VITE_AUTH0_AUDIENCE=AUTH0-AUDIENCE
