+++ package.json
@@ -32,6 +32,8 @@
     "styled-components": "^5.3.11"
   },
   "devDependencies": {
+    "@swc/cli": "^0.1.62",
+    "@swc/core": "^1.3.76",
     "@types/auth0-lock": "^11.27.5",
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
