+++ server/.env.example
@@ -1,5 +1,9 @@
-PORT=8000
 CLIENT_ORIGIN_URL=http://localhost:5173
-RELOAD=True
+
 AUTH0_DOMAIN=AUTH0-DOMAIN
 AUTH0_AUDIENCE=AUTH0-AUDIENCE
+
+AWS_KEY=AWS-IAM-KEY
+AWS_SECRET=AWS-IAM-SECRET
+AWS_REGION=ap-northeast-1
+STAGE=local
