+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import Router from './pages/router';
-import { Auth0ProviderWithNavigate } from './utils/auth0';
+import { Auth0ProviderWithNavigate } from './libs/auth0';
 
 const App = () => {
   return (
